# design/openapi.yaml
openapi: 3.1.0
info:
  title: IdeaBank
  description: IdeaBank
  version: 1.0.0
servers:
  - url: 'http://127.0.0.1:8080'
paths:
...
   /v1/api/chat/{idea_id}/create:
     post:
       tags:
         - chats
       summary: Создание чата.
       description: Создание чата по конкретной идеи. 1 идея - 1 чат
       parameters:
         - name: idea_id
           in: path
           required: true
           schema:
             type: integer
       security:
         - AccessToken: []
       responses:
         '200':
           description: Успешное создание чата для идеи
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/create-chat'
         '404':
           description: Если идея не найдена
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   detail:
                     type: string
                     example: No idea found
         '403':
           description: | 
                  Если учетная запись пользователя не активная
                  или работа над идеей не начата.
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/forbidden'
         '401':
           description: При закончившемся или некорректном refresh_token
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/bad-refresh'
         '409':
           description: Если чат уже существует
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/conflict'
   /v1/api/chat/:
     get:
       tags:
         - chats
       summary: Получение чата/-ов
       description: | 
              При указании параметра id получается информация         
              по конкретному чату, без указания по всем
              чатам где состоит пользователь
       parameters:
         - name: id
           in: query
           required: false
           schema:
             type: integer
       security:
         - AccessToken: []
       responses:
         '200':
           description: Успешное получение списка чата/-ов в которых состоит пользователь
           content:
             application/json:
               schema:
                 type: array
                 items:
                   $ref: '#/components/schemas/get-chat'

         '404':
           description: Если чат не найден
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   detail:
                     type: string
                     example: No idea found
         '403':
           description: Если учетная запись пользователя не активная
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/forbidden'
         '401':
           description: При закончившемся или некорректном refresh_token
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/bad-refresh'
   /v1/api/chat/{chat_id}/close:
     post:
       summary: Закрытие чата
       description: Любой участник чата может его закрыть
       parameters:
         - name: chat_id
           in: path
           required: true
           schema:
             type: integer
       security:
         - AccessToken: []
       responses:
         '200':
           description: При успешном закрытии чата
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   chat_id:
                     type: integer
                     example: 12
         '404':
           description: Если чат не найден
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   detail:
                     type: string
                     example: No idea found
         '403':
           description: | 
                  Если учетная запись пользователя не 
                  активная или пользователь не участник чата
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/forbidden'
   /v1/api/chat/{chat_id}/message/fetch:
     get:
       tags:
         - chats
       summary: Получение сообщений
       description: Чтение сообщений из чата 
       parameters:
         - name: chat_id
           in: path
           required: true
           schema:
             type: integer
       security:
         - AccessToken: []
       responses:
         '200':
           description: Список идей
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/get-messages'
         '404':
           description: Если чат не найден
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   detail:
                     type: string
                     example: No idea found
         '403':
           description: | 
                  Если учетная запись пользователя не 
                  активная или пользователь не участник чата
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/forbidden'
   /v1/api/chat/{chat_id}/message/send:
     post:
       tags:
         - chats
       summary: Отправка сообщений
       description: Отправка сообщений в чата
       parameters:
         - name: chat_id
           in: path
           required: true
           schema:
             type: integer
       requestBody:
         required: true
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/send-message'
       security:
         - AccessToken: []
       responses:
         '200':
           description: Успешная отправка
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   message_id:
                     type: integer
                     example: 20
         '403':
           description: Если учетная запись пользователя не активная
           content:
             application/json:
               schema:
                 $ref: '#/components/responses/forbidden'
         '400':
           description: Если сообщение не отправлено.
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   detail:
                     type: integer
                     example: Bad request

  # КОМПОНЕНТЫ
components:
  schemas:
...
    get-chat:
      type: object
      properties:
        chat_id:
          type: integer
          example: 12
        idea_id:
          type: integer
          example: 20
        interlocutors_ids:
          type: array
          items:
            type: integer
          example: [1,2]
        chat_status:
          type: string
          example: active
        idea_status:
          type: string
          example: in-work
    send-message:
      type: object
      properties: 
        msg:
          type: string
          example: Hi!
  responses:
...
    forbidden:
      type: object
      properties:
        detail:
          type: string
          example: Forbidden
    conflict:
      type: object
      properties:
        detail:
          type: string
          example: Conflict
    bad-refresh:
      type: object
      properties:
        detail:
          type: string
          example: Invalid refresh token
...
    create-chat:
      type: object
      properties:
        chat_id:
          type: integer
          example: 505
    get-messages:
      type: object
      properties:
        msg:
          type: string
          example: Hello!
        its_your_msg:
          type: boolean
          example: false
  securitySchemes:
    AccessToken:
      type: apiKey
      in: cookie
      name: access_token

    RefreshToken:
      type: apiKey
      in: cookie
      name: refresh_token